requirements.txt
accelerate==0.18.0
aiohttp==3.8.3
aiosignal==1.2.0
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrdict==2.0.1
attrdict3==2.0.2
attridict==0.0.6
attrs==22.1.0
backcall==0.2.0
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
comm==0.1.2
contourpy==1.1.0
cryptography==38.0.4
cycler==0.11.0
datasets==2.12.0
debugpy==1.6.6
decorator==5.1.1
dill==0.3.6
docker-pycreds==0.4.0
docopt==0.6.2
executing==1.2.0
fastprogress==1.0.0
filelock==3.9.0
flit_core==3.6.0
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.15.1
idna==3.4
ipykernel==6.23.2
ipython==8.14.0
jedi==0.18.2
joblib==1.2.0
jsonlines==2.0.0
jupyter_client==8.0.3
jupyter_core==5.3.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.2
multiprocess==0.70.14
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==22.3.1
pipreqs==0.4.13
platformdirs==3.1.1
pooch==1.4.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
Pygments==2.14.0
pyOpenSSL==22.0.0
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
pyzmq==25.0.1
regex==2022.7.9
requests==2.28.1
responses==0.13.3
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
sentry-sdk==1.27.0
seqeval==1.2.2
setproctitle==1.3.2
setuptools==65.6.3
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
threadpoolctl==3.1.0
tokenizers==0.11.4
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.0.dev0
typing_extensions==4.4.0
urllib3==1.26.14
wandb==0.15.4
wcwidth==0.2.6
wheel==0.37.1
wordcloud==1.9.2
xxhash==2.0.2
yarg==0.1.9
yarl==1.8.1
